Simplify TypeScript configurations
Let most of the work be done by the default TypeScript compiler settings. Instead of having separate compiler settings for the scripts (ESNext) and the package (CommonJS), just rely on the recommended settings for the scripts and make the tsconfig.json the default one so it gets automatically used for the package.