Ignore files that are not JSON
While testing the code, I found that it choked trying to parse files that are not valid JSON. This can happen when you compress the file data using the Finder, for example. In that case, the workspace folder contains `.DS_Store` files, which you definitely want to ignore.

In any case, since the spec is pretty clear about the filetypes one can expect to find in this folder, I guess it doesn't hurt to filter out unexpected files.