Fixes a bug with using string.length to calculate bytesize
When using double-byte characters (like •), the length of the string we want to store won't match the bytes needed to store the string data on disk. We want to use a more reliable method to find the length, like `Buffer.byteLength(p)`, instead of the more brittle `p.length` we were using.